Starting UI Development...WIP
Starting to develop the UI for the various Windows and UserControls to
be used with a DockPanel interface.  The MySQL Connector for C# assembly
has been added as a reference to the project.  Downloading and
installing the connector is a requirement before it will likely execute.
It doesn't make use of the connector just yet so it "may" execute, but I
doubt it will build which should affect that.