UI Windows/Dialogs Updated
Most of the UI has been created for the application.  There will be at
least one window that needs to be created related to the scheduling of a
selected medication to be added to a patient for administration, and
possibly some other odds and ends, but most of it is done.