Added MySqlHelper class and login logic.
Added a MySqlHelper class in order to eliminate MySQL Connector
boilerplate code and implemented the login logic.  The admin root
account was created for the master account, and a secondary account
marked as account type station was added to test screen navigation
paths.  Logging in as admin loads a SimulationPoolControl as the
starting place, and logging in as a station loads a PatientPoolControl
as the starting place.